fixed bug in benchmark main

+++ src/clpp/clppProgram.h
@@ -58,7 +58,7 @@ protected:
 
 	static size_t toMultipleOf(size_t N, size_t base) 
 	{
-		return (ceil((double)N / (double)base) * base);
+		return static_cast<size_t>((ceil((double)N / (double)base) * base));
 	}
 };
